Adds passport and quard

+++ model/contacts.js
@@ -1,38 +1,37 @@
 const Contact = require('./schemas/contact')
 
-const listContacts = async () => {
-  const results = await Contact.find({})
+const listContacts = async (userId) => {
+  const results = await Contact.find({ owner: userId }).populate({
+    path: 'owner',
+    select: 'email -_id'
+  })
   return results
 }
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findOne({ _id: contactId })
+const getContactById = async (contactId, userId) => {
+  const result = await Contact.findOne({ _id: contactId, owner: userId }).populate({
+    path: 'owner',
+    select: 'email -_id'
+  })
   return result
 }
 
 const addContact = async (body) => {
-  const contactsList = await listContacts()
-  const isExist = contactsList.find(({ name }) => name === body.name)
-  if (!isExist) {
-    const result = Contact.create(body)
-    return result
-  } else {
-    const message = `${body.name} is already in your contacts list`
-    return message
-  }
+  const result = await Contact.create(body)
+  return result
 }
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, userId) => {
   const result = await Contact.findByIdAndUpdate(
-    { _id: contactId },
+    { _id: contactId, owner: userId },
     { ...body },
     { new: true }
   )
   return result
 }
 
-const removeContact = async (contactId) => {
-  const result = await Contact.findOneAndRemove({ _id: contactId })
+const removeContact = async (contactId, userId) => {
+  const result = await Contact.findOneAndRemove({ _id: contactId, owner: userId })
   return result
 }
